March updates rc3 (#96)
* make mlst formver a float

* remove mdb001

* Remove literal string None from invalid text

* Fix cls values to be floats

* Make affiliates integers as that's easier to work with

* Update NACCREFR to not let 9 override in I4

* simplify slightly

+++ nacc_attribute_deriver/attributes/derived/modules/mlst/form_mlst.py
@@ -43,7 +43,7 @@ class MilestoneAttributeCollection(AttributeCollection):
             msg = f"Current file is not a MLST form: found {module}"
             raise InvalidFieldError(msg)
 
-        self.__formver = self.__mlst.get_required("formver", int)
+        self.__formver = self.__mlst.get_required("formver", float)
 
     def get_date(self) -> Optional[date]:
         return self.__mlst.get_date()
